# BitmapByteQRCode performance optimization