--- a/QRCoder/BitmapByteQRCode.cs
+++ b/QRCoder/BitmapByteQRCode.cs
@@ -113,7 +113,7 @@ public class BitmapByteQRCode : AbstractQRCode, IDisposable
                 bmp.AddRange(padding);
             }
         }
-        
+
         return bmp.ToArray();
     }