@codebude 2024-06-22
@Shane32 May I ask you for a review? (Please be critical with the benchmark project. I tried to exclude the QRCode data generation, as it is ramp-up stuff, from the benchmarks, but I'm not sure, if I did it right.)