@codebude 2024-06-23
> Looks great! Benchmarks look great! I've just jotted down some notes for additional improvements that could be made, whether that be within this PR or another one.

Hi Shane, thanks for your review. I implemented and tested all of your suggestions. Thanks for the great input! :-)
The results have improved once again. (I have inserted the new benchmarks in the first post above).